requirements.txt
# Core ML
torch>=2.1
torchvision
torchaudio

# NLP & Transformers (updated for Python 3.13 compatibility)
transformers>=4.40
sentence-transformers>=2.7
spacy>=3.7
nltk>=3.8

# ML & Data
scikit-learn>=1.4
pandas>=2.2
numpy>=1.26
imbalanced-learn>=0.12

# Explainability
shap>=0.45

# Vector Database
chromadb>=0.5

# Experiment Tracking
mlflow>=2.12

# Visualization
matplotlib>=3.8
seaborn>=0.13
plotly>=5.20

# Web App
streamlit>=1.33

# Testing
pytest>=8.0

# Utilities
pyyaml>=6.0
python-dotenv>=1.0
tqdm>=4.66